+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 [Upgrade Instructions](https://maru.readme.io/v0.12/docs/upgrade-instructions-from-v011) From v0.11 to v0.12
 
-## v0.12.4-dev
+## v0.12.4 (2017-10-17)
 * Enhancements
   * make regexp validator support list
   * support mount alias modules